Fixed an error relating to rumbe states
Fixed an being raied when rumble state changed by external library stopping the inputs library from being able to read the state of the controller. If this is not changed, the library instead throws an error when reading the controller state using get_gamepad() after the rumble state has been changed.